+++ apps/website/components/footer.tsx
@@ -1,13 +1,13 @@
-import { useTranslations } from 'next-intl';
-import { useRouter } from 'next/router';
+import { useTranslations } from "next-intl";
+import { useRouter } from "next/router";
 
-import { IGeneralSettings } from '../lib/general-settings';
-import socialMedia from '../lib/models/socialMedia';
-import { IPartner } from '../lib/partner';
-import NewsletterRegistration from './newsletterRegistration';
-import Partner from './partner';
-import Bubble from './shared/bubble';
-import Link from './shared/link';
+import { IGeneralSettings } from "../lib/general-settings";
+import socialMedia from "../lib/models/social-media";
+import { IPartner } from "../lib/partner";
+import NewsletterRegistration from "./newsletterRegistration";
+import Partner from "./partner";
+import Bubble from "./shared/bubble";
+import Link from "./shared/link";
 
 interface FooterProps {
   sponsorList: IPartner[];
@@ -30,14 +30,14 @@ const Footer = ({
   const difference = startDate.getTime() - new Date().getTime();
   let days = Math.ceil(difference / (1000 * 3600 * 24));
   days = days < 0 ? 0 : days;
-  const startAndEndDateString = `${startDate.toLocaleDateString(router.locale, {day: "numeric", month: "numeric"})} - ${endDate.toLocaleDateString(router.locale)}`;
+  const startAndEndDateString = `${startDate.toLocaleDateString(router.locale, { day: "numeric", month: "numeric" })} - ${endDate.toLocaleDateString(router.locale)}`;
 
   return (
     <>
       <div className="mt-8 sm:mt-12">
         <NewsletterRegistration />
       </div>
-      <div className="flex flex-wrap items-center justify-center gap-4 mt-8">
+      <div className="mt-8 flex flex-wrap items-center justify-center gap-4">
         {socialMedia.map((element, index) => (
           <Link href={element.url} key={index} className="hover:no-underline">
             <Bubble>
@@ -46,7 +46,7 @@ const Footer = ({
           </Link>
         ))}
       </div>
-      <div className="grid grid-cols-1 gap-4 px-3 mt-10 lg:grid-cols-3 sm:gap-6">
+      <div className="mt-10 grid grid-cols-1 gap-4 px-3 sm:gap-6 lg:grid-cols-3">
         <Partner label={t("sponsors").toString()} list={sponsorList} />
         <Partner
           label={t("media-partners").toString()}
@@ -54,7 +54,7 @@ const Footer = ({
         />
         <Partner label={t("supported-by").toString()} list={additionalList} />
       </div>
-      <div className="w-full max-w-3xl px-3 pb-12 mx-auto mt-8 text-center sm:pb-20 font-content">
+      <div className="mx-auto mt-8 w-full max-w-3xl px-3 pb-12 text-center font-content sm:pb-20">
         <p>
           Immergut Festival / Am Bürgerseeweg 28 / 17235 Neustrelitz
           <br />
@@ -72,9 +72,11 @@ const Footer = ({
           </Link>
         </p>
       </div>
-      {generalSettings.countdown.showCountdown && <div className="fixed bottom-0 flex justify-center w-full py-1 text-lg border-t-2 sm:text-4xl bg-secondary border-primary font-important">
-        {startAndEndDateString} ... noch {days} Tag{days !== 1 ? "e" : ""}!
-      </div>}
+      {generalSettings.countdown.showCountdown && (
+        <div className="fixed bottom-0 flex w-full justify-center border-t-2 border-primary bg-secondary py-1 font-important text-lg sm:text-4xl">
+          {startAndEndDateString} ... noch {days} Tag{days !== 1 ? "e" : ""}!
+        </div>
+      )}
     </>
   );
 };
