+++ apps/website/lib/shared/sanity-image-url.ts
@@ -4,7 +4,7 @@ import imageUrlBuilder from "@sanity/image-url";
 import { ImageUrlBuilder } from "@sanity/image-url/lib/types/builder";
 import { SanityImageSource } from "@sanity/image-url/lib/types/types";
 
-import { sanityClient } from "./sanityClient";
+import { sanityClient } from "./sanity-client";
 
 export interface SanityImage {
   hotspot: {
